未開放のマスのコンポーネント切り出し
